[console] - refactored new auth guards for cleanliness
