+++ console/src/auth/index.ts.ts
@@ -1,0 +1,1 @@
+export * as Auth from "@/auth/Guard";
